# backend/docs/swagger.yaml
definitions:
...
  ErrorResponseModel:
    properties:
      error:
        type: string
    type: object
...
  MaterialCreate:
    properties:
      name:
        type: string
    type: object
  MaterialObject:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  SystemsReadObject:
    properties:
      id:
        type: string
      material_name:
        type: string
      name:
        type: string
    type: object
  SystemsWriteObject:
    properties:
      material_name:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /materials:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all materials from the system
      produces:
      - application/json
      responses:
        "200":
          description: List of all materials
          schema:
            items:
              $ref: '#/definitions/MaterialObject'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get all materials
      tags:
      - materials
    post:
      consumes:
      - application/json
      description: Add a new material to the system with the provided details
      parameters:
      - description: Material details
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/MaterialCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Material created successfully
          schema:
            type: string
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Create a new material
      tags:
      - materials
  /materials/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a material from the system using its unique identifier
      parameters:
      - description: ID of the material to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material deleted successfully
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Delete a material
      tags:
      - materials
    get:
      consumes:
      - application/json
      description: Retrieve a specific material by its unique identifier
      parameters:
      - description: ID of the material to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material details
          schema:
            $ref: '#/definitions/MaterialObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get material by ID
      tags:
      - materials
  /systems:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all systems from the database
      produces:
      - application/json
      responses:
        "200":
          description: List of all systems
          schema:
            items:
              $ref: '#/definitions/SystemsReadObject'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get all systems
      tags:
      - systems
    post:
      consumes:
      - application/json
      description: Add a new system to the database with the provided details
      parameters:
      - description: System details
        in: body
        name: system
        required: true
        schema:
          $ref: '#/definitions/SystemsWriteObject'
      produces:
      - application/json
      responses:
        "201":
          description: System created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Create a new system
      tags:
      - systems
  /systems/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a system from the database using its unique identifier
      parameters:
      - description: ID of the system to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: System deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Delete a system
      tags:
      - systems
    get:
      consumes:
      - application/json
      description: Retrieve a specific system by its unique identifier
      parameters:
      - description: ID of the system to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: System details
          schema:
            $ref: '#/definitions/SystemsReadObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Get system by ID
      tags:
      - systems
    put:
      consumes:
      - application/json
      description: Update an existing system's information by ID
      parameters:
      - description: ID of the system to update
        in: path
        name: id
        required: true
        type: string
      - description: Updated system details
        in: body
        name: system
        required: true
        schema:
          $ref: '#/definitions/SystemsWriteObject'
      produces:
      - application/json
      responses:
        "200":
          description: System updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponseModel'
      summary: Update system
      tags:
      - systems
...
swagger: "2.0"